refactor: health bar
