+++ Assets/Scripts/Interfaces/Base.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ebe04fd599a1742b793dca4fcee82cc0
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
