+++ Assets/Scripts/Interfaces/Base/BaseHealthBar.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a02005f013cd944cf9bf36f174b85c99
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
